[BeamPlastic] CONVERT raw pointers into smart pointers
In the BeamPlasticFEMForceField, m_ConstitutiveLaw, and m_edgeHandler
are changed from raw pointers to std::unique_ptr. This allows to reduce
the definition of the destructor, which is set to default directly in
the header.
m_topology is kept as a raw pointer, as it is not allocated in the
component, but retrieved from the context.