[BeamPlastic] Improve double threshold for stress computation
The use of plasticity requires to measure the norm of stress tensors and
compare it to 0. Until now, we used an empirical fixed threshold,
lacking genericity, and potentially misleading as the value of the
stress tensor may vary greatly depending on the material deformation.
To improve this, we now used a 'fixed' threshold computed both from the
std double espilon and the material yield stress, which is typically in
the same order as the stress levels where plasticity occurs (for which
the comparison of the stress tensor norm to 0 needs to be precise).