[BeamPlastic] Implements quadrature methods
Following the addition of Gauss point and Interval integration
structures, we add methods to compute them for the beam elements, and to
perform Gaussian integration. At the moment, these elements are left
unused, but they should replace the mechanical data in BeamInfo as well
as the Gaussian.h and Quadrature.h files in future commits.